# webservice/src/main/resources/swagger/obm-api.yaml
swagger: "2.0"
info:
  description: "Service for OrchardBroadcastMonitor"
  version: "1.0.0"
  title: "OrchardBroadcastMonitorApi"
tags:
  - name: "OBM"
    externalDocs:
      description: "Find out more"
      url: "https://github.com/adamsiedlecki/OBM"
schemes:
  - "https"
  - "http"
paths:
  /broadcasts:
    post:
      tags:
        - "OBM"
      summary: "Processes broadcast information"
      operationId: "processBroadcastInformation"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "BroadcastInfoInput"
          description: Body of the broadcast infotmation request
          required: true
          schema:
            $ref: "#/definitions/BroadcastInfoInput"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request, text property is required!"
        "500":
          description: "Unexpected error."
    get:
      tags:
        - "OBM"
      summary: "Gives recent broadcasts"
      operationId: "getBroadcastInformation"
      produces:
        - "application/json"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/BroadcastListOutput"
        "400":
          description: "Bad request, text property is required!"
        "500":
          description: "Unexpected error."

definitions:
  BroadcastInfoInput:
    type: "object"
    properties:
      rssi:
        description: "Received Signal Strength Indication"
        type: integer
      text:
        description: "Text representation of the broadcast encoded in base64"
        type: string
    required:
      - text

  BroadcastListOutput:
    type: "object"
    properties:
      bList:
        description: "Received Signal Strength Indication"
        type: array
        items:
          $ref: "#/definitions/Broadcast"
    required:
      - bList

  Broadcast:
    type: "object"
    properties:
      text:
        type: string
      dateTime:
        type: string
        format: date-time
    required:
      - text
      - dateTime